Cleaned up the code in the ground intake folder and wrist folder which was removed and cleaned up auton commands
