+++ src/main/java/frc/robot/subsystems/groundIntake/RunGroundIntakeWithJoystick.java
@@ -9,39 +9,58 @@ import edu.wpi.first.wpilibj2.command.CommandBase;
 
 public class RunGroundIntakeWithJoystick extends CommandBase {
   /** Creates a new RunGroundIntakeWithJoystick. */
-  private GroundIntake groundIntake;
+  // private GroundIntake groundIntake;
   private XboxController joystick;
-  public RunGroundIntakeWithJoystick(GroundIntake groundIntake, XboxController joystick) {
-    this.groundIntake = groundIntake;
+  private GroundIntakeSubsystem groundIntakeSubsystem;
+
+  public RunGroundIntakeWithJoystick(GroundIntakeSubsystem groundIntakeSubsystem, XboxController joystick) {
+    // this.groundIntake = groundIntake;
     this.joystick = joystick;
-    addRequirements(groundIntake);
+    this.groundIntakeSubsystem = groundIntakeSubsystem;
+    addRequirements(groundIntakeSubsystem);
     // Use addRequirements() here to declare subsystem dependencies.
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    groundIntake.coastWrist();
+    groundIntakeSubsystem.wristCoast();
+    groundIntakeSubsystem.setWristPower(0);
+    groundIntakeSubsystem.intakeCoast();
+    groundIntakeSubsystem.setIntakeSpeed(0);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
     if (Math.abs(joystick.getRightTriggerAxis())  > 0.01 ) {
-      groundIntake.setPowerWrist(joystick.getRightTriggerAxis());
+      groundIntakeSubsystem.setIntakeSpeed(joystick.getRightTriggerAxis()/5);
     }  
     if (-(joystick.getLeftTriggerAxis())  < -0.01 ) {
-      groundIntake.setPowerWrist(-joystick.getLeftTriggerAxis());
-    }  
+      groundIntakeSubsystem.setIntakeSpeed(-joystick.getLeftTriggerAxis()/5);
+    }
+
+    if(joystick.getRightY()>0.07){
+      groundIntakeSubsystem.wristCoast();
+      groundIntakeSubsystem.setWristPower(joystick.getRightY()/4);
+    }
+    else {
+      groundIntakeSubsystem.wristBrake();
+      groundIntakeSubsystem.setWristPower(0);
+    }
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    groundIntakeSubsystem.wristBrake();
+    groundIntakeSubsystem.setWristPower(0);
+    groundIntakeSubsystem.setIntakeSpeed(0);
+  }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return false;
+    return groundIntakeSubsystem.isWristOverheating();
   }
 }
